+++ CHANGELOG.rst
@@ -1,6 +1,11 @@
 Changelog
 #########
 
+*latest*
+--------
+
+- ``printinfo``: Print MKL-info (if available) independently of ``numexpr``.
+
 v1.7.0 - *2018-05-23*
 ---------------------